Initial version of copilot-bluespec
This is an implementation of a Copilot Bluespec backend, based off of an
original prototype written by Ivan Perez. The design of `copilot-bluespec` is
described in more detail in `DESIGN.md`. The code that it generates is very
similar to the style of code that `copilot-c99` generates: each stream is
represented as a ring buffer with a corresponding index tracking the latest
entry in the buffer, and invocation of the `step` function (or, in the case of
`copilot-bluespec`, the `step` rule) computes a new entry and increments the
index.

I've added unit tests (again, inspired by `copilot-bluespec`) that test all of
the operations that `copilot-bluespec` currently supports. The operations that
`copilot-bluespec` supports are largely determined by what the Bluespec
language itself supports; see the `DESIGN.md` document for more discussion on
this point.

Co-authored-by: Ivan Perez <ivan.perezdominguez@nasa.gov>