Test suite: wrap some properties in `once`
Prior to `QuickCheck-2.15`, properties without any quantification (e.g.,
`quickCheck True`) would only be run a single time. This is no longer the case
in `QuickCheck-2.15` or later, as these properties will now be run multiple
times, just like properties with arguments.

We use `ioProperty` to define a small number properties without any arguments
in the `copilot-bluespec` test suite. Unfortunately, this means that we were
previously compiling and simulating, say, 7 Bluespec programs prior to
`QuickCheck-2.15`, but we now we are compiling and simulating _700_ Bluespec
programs after `QuickCheck-2.15`. This means that it takes an extraordinarily
long time to run the test suite using `QuickCheck-2.15`.

This patch restores the old behavior of pre-2.15 versions of `QuickCheck` by
wrapping these properties in `once`.