Translate array updates to Bluespec. Refs #19.
This adds the necessary code to translate Copilot expressions that update an
element of an array to Bluespec, via Bluespec's `update` function (for function
array updates). This leverages the same logic for initializing vectors one
element at a time using `update`, so this commit factors out the code that is
shared in common.