Implement missing/buggy FP operations with `import "BDPI"`. Refs #28.
This implements `exp`, `log`, etc. (which were previously missing in
`copilot-bluespec`) as well as `sqrt` (which was previously buggy) using
Bluespec's `import "BDPI"` feature, allowing these operations to be implemented
in terms of C code. While not the most performant implementations around, these
are much more likely to be correct than hand-written code. We can work on
improving the performance of these operations in future work.