Expose Copilot.Compile.Bluespec.External. Refs #36.
The non-public `Copilot.Compiler.Bluespec.External` module lives under a
`shared/` directory so that it be imported by both the library code and the
test suite code. This is a questionable design, as if the test suite needs to
use the code, then it should be exposed by the library.

This commit moves `Copilot.Compiler.Bluespec.External` from `shared/` to
`src/`, exposes the module as part of the public API, and removes the `shared/`
directory. As a result, the test suite can now import
`Copilot.Compiler.Bluespec.External` like it does other parts of the
`copilot-bluespec` API.