Typofix: Bluespec, not C99. Refs #36.
The non-public `Copilot.Compiler.Bluespec.External` module lives under a
`shared/` directory so that it be imported by both the library code and the
test suite code. This is a questionable design, as it means that the code for
the library is not self-contained in a single `src/` directory. (There exists a
similar problem for the test suite.)

The comments in `Copilot.Compiler.Bluespec.External` mistakenly reference the
C99 backend instead of the Bluespec backend, which will make it confusing for
users when this module is moved to a more prominent, Bluespec-specific
location. This commit fixes the comments to reference the Bluespec backend
instead.