Typofix: Bluespec, not C99. Refs #36.
The non-public `Copilot.Compiler.Bluespec.External` module lives under a
`shared/` directory so that it be imported by both the library code and the
test suite code. This is a questionable design, as if the test suite needs to
use the code, then it should be exposed by the library.

The comments in `Copilot.Compiler.Bluespec.External` mistakenly reference the
C99 backend instead of the Bluespec backend, which will make it confusing for
users when this module is exposed as part of the public interface. This commit
fixes the comments to reference the Bluespec backend instead.