Move Copilot.Compile.Bluespec.External out of shared directory. Refs #36.
The non-public `Copilot.Compiler.Bluespec.External` module lives under a
`shared/` directory so that it be imported by both the library code and the
test suite code. This is a questionable design, as it means that the code for
the library is not self-contained in a single `src/` directory. (There exists a
similar problem for the test suite.)

This commit moves `Copilot.Compiler.Bluespec.External` from `shared/` to
`src/`. It also creates a duplicate version of the module under the name
`Test.Copilot.Compiler.Bluespec.External`. This code duplication is not ideal,
but it is better than the previous status quo. The task of de-duplicating it
will be deferred to a separate issue
(https://github.com/Copilot-Language/copilot-bluespec/issues/3).